Improve user experience
Added confirmation on account deletion to prevent deleting account accidentally.

Improved minor panel visuals to enhance appearance.

Fixed bug where all buttons to other tasks disappear when a task is being modified.

+++ client/src/pages/settings/index.js
@@ -67,16 +67,20 @@ const Settings = () => {
     };
 
     const removeUser = (imageId) => {
-        deleteUser()
-        .then(() => {
-            deleteImage(imageId)
-            .then((navigate) => {
-                if (navigate) {
-                    authReset();
-                    UserRedirect('/auth');
-                }
+        const confirm = window.confirm('Delete account?');
+
+        if (confirm === true) {
+            deleteUser()
+            .then(() => {
+                deleteImage(imageId)
+                .then((navigate) => {
+                    if (navigate) {
+                        authReset();
+                        UserRedirect('/auth');
+                    }
+                });
             });
-        });
+        }
     };
 
     return (
